web/requirements.txt
coverage==4.2
Django==1.10.2
django-appconf==1.0.2
django-authtools==1.5.0
django-autocomplete-light==3.2.1
django-braces==1.10.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-extra-views==0.8.0
django-model-utils==2.6
django-notifications-hq==1.2
jsonfield==1.0.3
Pillow==6.1.0
pytz==2016.7
six==1.10.0
sqlparse==0.2.2
psycopg2-binary

